init_refa_footer
